+++ test/acceptance/batch/batch-drain.test.js
@@ -5,6 +5,7 @@ var assert = require('../../support/assert');
 var redisUtils = require('../../support/redis_utils');
 var batchFactory = require('../../../batch/index');
 
+var BatchLogger = require('../../../batch/batch-logger');
 var JobPublisher = require('../../../batch/pubsub/job-publisher');
 var JobQueue = require('../../../batch/job_queue');
 var JobBackend = require('../../../batch/job_backend');
@@ -16,11 +17,12 @@ describe('batch module', function() {
     var dbInstance = 'localhost';
     var username = 'vizzuality';
     var pool = redisUtils.getPool();
+    var logger = new BatchLogger(null, 'batch-queries');
     var jobPublisher = new JobPublisher(pool);
-    var jobQueue =  new JobQueue(metadataBackend, jobPublisher);
-    var jobBackend = new JobBackend(metadataBackend, jobQueue);
+    var jobQueue =  new JobQueue(metadataBackend, jobPublisher, logger);
+    var jobBackend = new JobBackend(metadataBackend, jobQueue, logger);
     var jobCanceller = new JobCanceller();
-    var jobService = new JobService(jobBackend, jobCanceller);
+    var jobService = new JobService(jobBackend, jobCanceller, logger);
 
     before(function (done) {
         this.batch = batchFactory(metadataBackend, pool);